Move to tests/conftest
