+++ conftest.py
@@ -1,5 +1,4 @@
 from typing import List
-from pydispatch import dispatcher
 import pytest
 import _pytest
 
@@ -18,17 +17,3 @@ def pytest_collection_modifyitems(config: _pytest.config.Config,
     for item in items:
         if "slow" in item.keywords:
             item.add_marker(skip_slow)
-
-
-@pytest.fixture(autouse=True)
-def clean_dispatcher():
-    """
-    Dispatcher is a global object shared between different tests so it may
-    happen that one tests subscribes to a signal than some completely different
-    tests sends this signal triggering code from the first test which is
-    completely unexpected and undefined.
-    """
-    yield
-    dispatcher.connections = {}
-    dispatcher.senders = {}
-    dispatcher.sendersBack = {}
